Check for writable gfile which has no p file,  abort  "bk mv", "bk rm",  "bk mvdir", "bk rmdir" if found..
bk: 39ac464bppDzh4OXnEh1OK_oTo-ypw

+++ src/sccs_mv.c
@@ -40,6 +40,13 @@ sccs_mv(char *name, char *dest, int isDir, int isDelete)
 		return (1);
 	}
 
+	if (!HAS_PFILE(s) && S_ISREG(s->mode) && IS_WRITABLE(s)) {
+		fprintf(stderr,
+		    "sccsmv: %s is writable but not edited\n",
+		    s->gfile);
+		return (1);
+	}
+
 	/* XXX - shouldn't this call sccs_clean()? */
 	if (HAS_PFILE(s) && !HAS_GFILE(s)){
 		unlink(s->pfile);
